requirements.txt
Django==1.4.5
MySQL-python==1.2.4
PIL==1.1.7
South==0.7.6
amqp==1.0.11
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.28
celery==3.0.19
distribute==0.6.36
django-braces==1.0.0
django-celery==3.0.17
django-disqus==0.4.1
django-extensions==1.0.3
django-guardian
django-ratings==0.3.7
django-registration==0.8
django-widget-tweaks==1.1.2
gunicorn==0.17.2
kombu==2.5.10
python-dateutil==2.1
pytz==2013b
six==1.3.0
wsgiref==0.1.2
wsgiref==0.1.2
Markdown==2.3.1
